Move useMemo in a separate hook
